Made a javascript to typescript refactor
