+++ packages/css-doc-page/README.md
@@ -48,7 +48,7 @@ Both are React components with illustrated in the next image:
 To create the documentation of a new component follow the next steps:
 
 1. Create a Markdown file for your component in `packages/css-doc-page/src/assets/markdown/{component}/file.md`.
-2. Add a view in a new folder called `packages/css-doc-page/src/views/{componentDoc}` with the next files: `componentDoc.jsx`, `componentDoc.scss` and `index.js`.
+2. Add a view in a new folder called `packages/css-doc-page/src/views/{componentDoc}` with the next files: `componentDoc.jsx`, `componentDoc.scss` and `index.tsx`.
 3. In the `componentDoc.jsx` follow the `<Hero/> + <Markdown/>` structure.
 4. Import the markdown files of step one, and use them as a prop of the `<Markdown/>` component.
 5. Associate the view path in the `packages/css-doc-page/src/containers/sidebar/Sidebar.jsx`
