+++ packages/sandbox/package.json
@@ -15,7 +15,7 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "serve": "node runServer.js",
+    "serve": "node runServer.ts",
     "serve:stop": "pm2 stop sandbox"
   },
   "eslintConfig": {
@@ -34,6 +34,10 @@
     ]
   },
   "devDependencies": {
+    "@types/react": "^17.0.40",
+    "@types/react-dom": "^17.0.13",
+    "@types/styled-components": "^5.1.24",
+    "typescript": "^4.6.2",
     "express": "^4.17.1",
     "pm2": "^4.5.4"
   }
